[refactoring] RoomService 분리
